Separate file for nms
